Update index_alpha.html
Fixed a bunch of bugs with regards to linking. The original program listed children, I changed it to parent so multiple items could be hidden inside one parent. This was okay, except I also added a function to show what items were in the room in addition to the room description. I exclude child objects, so I went through and reassigned children to their appropriate parents.